Splitting graphics into two different pages
